% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beep_success_error.R
\name{beep_success_error}
\alias{beep_success_error}
\title{beep on success or error}
\usage{
beep_success_error(expr, success = 4, error = 2)
}
\arguments{
\item{expr}{the expression to evaluate}

\item{success}{which beep to use upon success}

\item{error}{which beep to use upon an error happening}
}
\value{
whatever should have been returned
}
\description{
Given an expression to evaluate, beeps on error or success so you still know it is done.
}
