% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics_helpers.R
\name{ft_add_figure_numbers}
\alias{ft_add_figure_numbers}
\title{add numbers to figure files}
\usage{
ft_add_figure_numbers(
  figure_directory = ".",
  start_count = 1,
  count_location = "prefix",
  skip_numbers = NULL,
  regexp = "jpg$|png$|svg$"
)
}
\arguments{
\item{figure_directory}{the directory containing figure files}

\item{start_count}{what number starts the counting}

\item{count_location}{should the count be a \code{prefix} or \code{suffix}?}

\item{skip_numbers}{should any numbers be skipped due to inserting from a different location?}

\item{regexp}{what regex should be used to grab files}
}
\description{
Given that \code{rmd} and \code{qmd} chunks are \strong{normally} run sequentially, figures
should be generated in order that they are presented in the document. Therefore,
we should be able to roughly match them up with numbering generated by cross-referencing
using the modified times.
}
\seealso{
Other {Graphics}:
\code{\link[=ft_add_padding]{ft_add_padding()}},
\code{\link[=ft_discretize_colorscale]{ft_discretize_colorscale()}},
\code{\link[=ft_ggplot2_add_ellipse]{ft_ggplot2_add_ellipse()}},
\code{\link[=ft_ggplot2_change_zero]{ft_ggplot2_change_zero()}},
\code{\link[=ft_ggplot2_colors]{ft_ggplot2_colors()}},
\code{\link[=ft_ggplot2_legend_inside]{ft_ggplot2_legend_inside()}},
\code{\link[=ft_ggplot2_rotate_axis_labels]{ft_ggplot2_rotate_axis_labels()}},
\code{\link[=ft_ggplot2_self_colors]{ft_ggplot2_self_colors()}},
\code{\link[=ft_plot_temp_file]{ft_plot_temp_file()}},
\code{\link[=ft_upset_add_counts]{ft_upset_add_counts()}},
\code{\link[=ft_upset_comb_matrix]{ft_upset_comb_matrix()}}

Other {File Manipulation}:
\code{\link[=ft_rename_code]{ft_rename_code()}},
\code{\link[=ft_rename_outputs]{ft_rename_outputs()}},
\code{\link[=ft_zip_directory]{ft_zip_directory()}}
}
\concept{{File Manipulation}}
\concept{{Graphics}}
