% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics_helpers.R
\name{ft_complexheatmap_scale}
\alias{ft_complexheatmap_scale}
\title{generate ComplexHeatmap color scale}
\usage{
ft_complexheatmap_scale()
}
\value{
string
}
\description{
The code for creating a function that maps values to colors for \code{ComplexHeatmap}
heatmaps is not easy for me to remember. Here it is.
}
\details{
My favorite colors to use include:
\itemize{
\item \code{viridis::viridis}
\item \code{scico::scico(n, palette = "vik")}
}

You can also use \code{scico::scico_palette_show()} to see other colorblind safe palettes.
}
\examples{
ft_complexheatmap_scale()
}
\seealso{
Other {Code Generators}:
\code{\link[=ft_add_padding]{ft_add_padding()}},
\code{\link[=ft_data_dictionary]{ft_data_dictionary()}},
\code{\link[=ft_ggplot2_add_ellipse]{ft_ggplot2_add_ellipse()}},
\code{\link[=ft_ggplot2_change_zero]{ft_ggplot2_change_zero()}},
\code{\link[=ft_ggplot2_colors]{ft_ggplot2_colors()}},
\code{\link[=ft_ggplot2_legend_inside]{ft_ggplot2_legend_inside()}},
\code{\link[=ft_ggplot2_rotate_axis_labels]{ft_ggplot2_rotate_axis_labels()}},
\code{\link[=ft_ggplot2_self_colors]{ft_ggplot2_self_colors()}},
\code{\link[=ft_knitr_opts]{ft_knitr_opts()}},
\code{\link[=ft_plot_includes]{ft_plot_includes()}},
\code{\link[=ft_remove_figures]{ft_remove_figures()}},
\code{\link[=ft_rename_code]{ft_rename_code()}},
\code{\link[=ft_title_block]{ft_title_block()}},
\code{\link[=ft_upset_add_counts]{ft_upset_add_counts()}},
\code{\link[=ft_upset_comb_matrix]{ft_upset_comb_matrix()}}

Other {Colors}:
\code{\link[=ft_discretize_colorscale]{ft_discretize_colorscale()}},
\code{\link[=ft_ggplot2_colors]{ft_ggplot2_colors()}}
}
\concept{{Code Generators}}
\concept{{Colors}}
\concept{{Graphics} {ComplexHeatmap}}
