% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dive.R
\name{ft_dive}
\alias{ft_dive}
\title{dive into a dataframe}
\source{
Miles McBain \url{https://milesmcbain.micro.blog/2025/07/04/diveing-into-the-hunt-rstats.html}
}
\usage{
ft_dive(df)
}
\arguments{
\item{df}{the data.frame to investigate}
}
\description{
Drops the user into a debug session in the context of a given data.frame, which allows for calculations and
investigation of columns in a particular data.frame.
}
\examples{
\dontrun{
weird_df = data.frame(x = rnorm(200), id = sample(letters, 200, replace = TRUE))
weird_df$y = weird_df$x * 4
weird_df |>
  dplyr::filter(id \%in\% "a") |>
  ft_dive()
}



}
