% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_helpers.R
\name{ft_launch_watch_relaunch}
\alias{ft_launch_watch_relaunch}
\title{launch and relaunch a command}
\usage{
ft_launch_watch_relaunch(
  command,
  args = NULL,
  ntry = 10,
  check_interval = 60,
  status_expected = "running",
  ...
)
}
\arguments{
\item{command}{the command to run}

\item{args}{the command arguments}

\item{ntry}{how many times to try relaunching the command}

\item{check_interval}{how long to wait in seconds before checking on it}

\item{status_expected}{what status is expected, default is "running"}

\item{...}{other arguments passed to \code{sys::exec_background}}
}
\value{
NULL invisibly
}
\description{
Given a command string and arguments, runs it, checks if the process is still alive,
and if the process has died, launches it again. Requires packages \code{sys} and \code{ps}.
}
\seealso{
\code{\link[sys:exec_background]{sys::exec_background()}}, \code{\link[ps:ps_status]{ps::ps_status()}}

Other 'Analysis':
\code{\link[=ft_limma_code]{ft_limma_code()}},
\code{\link[=ft_set_function_params]{ft_set_function_params()}}
}
\concept{'Analysis'}
