% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_helpers.R
\name{ft_limma_code}
\alias{ft_limma_code}
\title{create limma code}
\usage{
ft_limma_code()
}
\value{
NULL, invisibly
}
\description{
A helper that provides a template for limma code to programmatically
set up the differences one wants to extract.
}
\examples{
ft_limma_code()

}
\seealso{
Other 'Analysis':
\code{\link[=ft_launch_watch_relaunch]{ft_launch_watch_relaunch()}},
\code{\link[=ft_set_function_params]{ft_set_function_params()}}
}
\concept{'Analysis'}
