% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beep_success_error.R
\name{ft_notify_success_error}
\alias{ft_notify_success_error}
\title{Notify on success or error}
\usage{
ft_notify_success_error(
  expr,
  success = "All done!",
  error = "Shit!",
  image = NULL
)
}
\arguments{
\item{expr}{the expression to run}

\item{success}{the message to show when no errors happen}

\item{error}{the message to show when an error is encountered}

\item{image}{an image to put on the notification}
}
\value{
the results of the expression, invisibly
}
\description{
Uses libnotify on Linux to give a system notification when sound nor \code{ntfy}
are available.
}
