% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quarto_create_post.R
\name{ft_quarto_create_post}
\alias{ft_quarto_create_post}
\title{create a new quarto blog post}
\usage{
ft_quarto_create_post(
  title,
  file = "index.qmd",
  subtitle = "",
  date = Sys.Date(),
  draft = FALSE,
  title_limit = 40,
  open_file = TRUE
)
}
\arguments{
\item{title}{the title of the post}

\item{file}{what file should be output (default is index.qmd)}

\item{subtitle}{what the subtitle should be}

\item{date}{the date of the post}

\item{draft}{is this a draft post?}

\item{title_limit}{maximum number of characters in the title}

\item{open_file}{should the file be opened by RStudio?}
}
\description{
Sets up the directory for a new blog post.
Note this function was modified from the one originally supplied
by \href{https://themockup.blog/posts/2022-11-08-use-r-to-generate-a-quarto-blogpost/}{Tom Mock}.
}
\seealso{
Other {Quarto}:
\code{\link[=ft_quarto_get_categories]{ft_quarto_get_categories()}}
}
\concept{{Quarto}}
