% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quarto_create_post.R
\name{ft_quarto_get_categories}
\alias{ft_quarto_get_categories}
\title{get categories from posts}
\usage{
ft_quarto_get_categories(post_directory = "posts")
}
\arguments{
\item{post_directory}{the directory that holds all of the posts}
}
\value{
tibble
}
\description{
Goes through the directory of quarto blog posts, and parses the yaml frontmatter
for the \code{categories} field, counts how many of each category there are and arranges
them in decreasing order as a {tibble}.
}
\seealso{
Other {Quarto}:
\code{\link[=ft_quarto_create_post]{ft_quarto_create_post()}}
}
\concept{{Quarto}}
