% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_helpers.R
\name{ft_r_launch_watch_relaunch}
\alias{ft_r_launch_watch_relaunch}
\title{launch and relaunch an R process}
\usage{
ft_r_launch_watch_relaunch(
  command,
  args = NULL,
  ntry = 10,
  check_interval = 60,
  status_expected = c("sleeping", "running"),
  ...
)
}
\arguments{
\item{command}{the R code to run as a string}

\item{args}{the command arguments}

\item{ntry}{how many times to try relaunching the command}

\item{check_interval}{how long to wait in seconds before checking on it}

\item{status_expected}{what status is expected, default is "sleeping" & "running"}

\item{...}{other arguments passed to \code{sys::exec_background}}
}
\value{
NULL invisibly
}
\description{
Given a bit of R code, runs it in the background, checks if the R process is still alive,
and if the R process has died, launches it again. Requires packages \code{sys} and \code{ps}.
}
\details{
Note: this writes the command provided to an R file in \code{tmp}, and then passes it via \code{std_in}.
}
\seealso{
\code{\link[sys:r_background]{sys::r_background()}}, \code{\link[ps:ps_status]{ps::ps_status()}}

Other 'Analysis':
\code{\link[=ft_launch_watch_relaunch]{ft_launch_watch_relaunch()}},
\code{\link[=ft_limma_code]{ft_limma_code()}},
\code{\link[=ft_set_function_params]{ft_set_function_params()}}
}
\concept{'Analysis'}
