% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knit.R
\name{ft_render_document}
\alias{ft_render_document}
\title{rmarkdown render}
\usage{
ft_render_document(
  document = NULL,
  working_directory = getwd(),
  format = "html_document",
  ...
)
}
\arguments{
\item{document}{the \code{Rmd} document being used.}

\item{working_directory}{the root directory of the \code{targets} project (default to getwd)}

\item{format}{the output format to generate (default is html_document)}

\item{...}{any other parameters used by \code{rmarkdown::render}}
}
\value{
the output of running \code{rmarkdown::render}
}
\description{
A version of \code{rmarkdown::render} that has sane defaults for when working in a \code{targets} project
where report documents are frequently kept in a subfolder, but \code{targets} needs to operate from
the root folder to find and load everything.
}
\seealso{
Other 'Document Rendering': 
\code{\link{ft_quarto_document}()}
}
\concept{'Document Rendering'}
