% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knit.R
\name{ft_rmarkdown_render}
\alias{ft_rmarkdown_render}
\title{rmarkdown render}
\usage{
ft_rmarkdown_render(document = NULL, working_directory = getwd(), ...)
}
\arguments{
\item{document}{the \code{Rmd} document being used.}

\item{working_directory}{the root directory of the \code{targets} project (default to getwd)}

\item{...}{any other parameters used by \code{rmarkdown::render}}
}
\value{
the output of running \code{rmarkdown::render}
}
\description{
A version of \code{rmarkdown::render} that has sane defaults for when working in a \code{targets} project
where report documents are frequently kept in a subfolder, but \code{targets} needs to operate from
the root folder to find and load everything, \strong{and} is run within a separate \code{callr} process so
all the outputs from the render are not in the local workspace.
}
\seealso{
Other 'Document Rendering': 
\code{\link{ft_quarto_render}()}
}
\concept{'Document Rendering'}
