% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timekeeping.R
\name{ft_tk_get_categories}
\alias{ft_tk_get_categories}
\title{get allowed timekeeping categories}
\usage{
ft_tk_get_categories()
}
\value{
string
}
\description{
get allowed timekeeping categories
}
\seealso{
Other {Timekeeping}:
\code{\link[=ft_setup_timekeeping]{ft_setup_timekeeping()}},
\code{\link[=ft_tk_check_log]{ft_tk_check_log()}},
\code{\link[=ft_tk_logfile]{ft_tk_logfile()}},
\code{\link[=ft_tk_logtask]{ft_tk_logtask()}},
\code{\link[=ft_tk_parselog]{ft_tk_parselog()}},
\code{\link[=ft_tk_removelast]{ft_tk_removelast()}},
\code{\link[=ft_tk_summarize_time]{ft_tk_summarize_time()}}
}
\concept{{Timekeeping}}
