% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timekeeping.R
\name{ft_tk_parselog}
\alias{ft_tk_parselog}
\title{Parse timekeeping log-file}
\usage{
ft_tk_parselog(log_file = ft_tk_logfile())
}
\arguments{
\item{log_file}{the logfile to use. default is \code{ft_tk_logfile}}
}
\value{
tibble of the log-file contents
}
\description{
Parse timekeeping log-file
}
\seealso{
Other {Timekeeping}:
\code{\link[=ft_setup_timekeeping]{ft_setup_timekeeping()}},
\code{\link[=ft_tk_check_log]{ft_tk_check_log()}},
\code{\link[=ft_tk_get_categories]{ft_tk_get_categories()}},
\code{\link[=ft_tk_logfile]{ft_tk_logfile()}},
\code{\link[=ft_tk_logtask]{ft_tk_logtask()}},
\code{\link[=ft_tk_removelast]{ft_tk_removelast()}},
\code{\link[=ft_tk_summarize_time]{ft_tk_summarize_time()}}
}
\concept{{Timekeeping}}
