% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/watch_process.R
\name{ft_watch_pid}
\alias{ft_watch_pid}
\title{watch pid}
\usage{
ft_watch_pid(
  pid = NULL,
  dead_action = function() {
     ft_notify_success_error("it died")
 },
  alive_action = NULL,
  check_interval = 60
)
}
\arguments{
\item{pid}{the process ID to watch}

\item{dead_action}{what to do if the process no longer exists (display message)}

\item{alive_action}{what to do if the process is alive}

\item{check_interval}{how often to check the process in seconds (default is 60)}
}
\description{
Given a process id (pid), watches the process, and if the process dies, then
takes an action given by the supplied function. Also able to take action if
the process is still alive.
}
\examples{
\dontrun{
use_pid = ps::ps_pid()
alive_action = function(){ft_notify_success_error(NULL, "it's running!")}
dead_action = function(){ft_notify_success_error(NULL, "it died")}
ft_watch_pid(use_pid, alive_action = alive_action, dead_action = dead_action)
}

}
\seealso{
\code{\link[=ft_notify_success_error]{ft_notify_success_error()}} \code{\link[=ft_ntfy_success_error]{ft_ntfy_success_error()}} \code{\link[=ft_beep_success_error]{ft_beep_success_error()}}
}
