% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename_outputs.R
\name{ft_zip_directory}
\alias{ft_zip_directory}
\title{zip a directory}
\usage{
ft_zip_directory(directory, zipfile, add_date = TRUE)
}
\arguments{
\item{directory}{what directory of files to zip}

\item{zipfile}{where to save the zip file}

\item{add_date}{should the date be added?}
}
\value{
path to file
}
\description{
Given a file folder, creates a zip file containing all the underlying files, starting
at the directory level.
}
\seealso{
Other {File Manipulation}:
\code{\link[=ft_add_figure_numbers]{ft_add_figure_numbers()}},
\code{\link[=ft_rename_code]{ft_rename_code()}},
\code{\link[=ft_rename_outputs]{ft_rename_outputs()}}
}
\concept{{File Manipulation}}
