% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beep_success_error.R
\name{notifier}
\alias{notifier}
\alias{ft_beep_success_error}
\alias{ft_ntfy_success_error}
\alias{ft_notify_success_error}
\title{Notifier functions}
\usage{
ft_beep_success_error(expr, success = 4, error = 2)

ft_ntfy_success_error(
  expr,
  success = "It Worked!",
  error = "Better Check on Me."
)

ft_notify_success_error(
  expr,
  success = "All done!",
  error = "Shit!",
  image = NULL
)
}
\arguments{
\item{expr}{the expression to evaluate}

\item{success}{what to use for success (sound or string)}

\item{error}{what to use for an error (sound or string)}

\item{image}{for \code{notify}, is there a specific image to use for the graphical notification?}
}
\value{
results of expression
}
\description{
These functions can be used to supply an alert to the user that the evaluation of
an expression succeeded or resulted in an error.
}
\details{
Each of these functions takes an expression as the first argument, so it is possible
to pipe into them, or just wrap the expression to run. They check if the evaluated
expression returns an error or not, so they can give the correct feedback to the user.
}
\examples{
if (require(beepr)) {
  # usual syntax, works
  ft_beep_success_error(Sys.sleep(2))

  # have an error (can't add a letter to a number)
  ft_beep_success_error(2 + "x")

  # pipe syntax
  Sys.sleep(2) |> ft_beep_success_error()
}

}
