% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beep_success_error.R
\name{ntfy_success_error}
\alias{ntfy_success_error}
\title{ntfy on success or error}
\usage{
ntfy_success_error(expr, success = "It Worked!", error = "Better Check on Me.")
}
\arguments{
\item{expr}{the expression to evaluate}

\item{success}{the message to send upon success}

\item{error}{the message to send upon an error}
}
\value{
whatever should have been returned
}
\description{
Given an expression to evaluate, sends a ntfy message on error or success so you still know it is done.
}
