% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename_outputs.R
\name{rename_outputs}
\alias{rename_outputs}
\title{rename outputs}
\usage{
rename_outputs(
  file_names = NULL,
  prefix = "",
  suffix = Sys.Date(),
  overwrite = FALSE
)
}
\arguments{
\item{file_names}{which files need to be renamed?}

\item{prefix}{what is the prefix to add?}

\item{suffix}{what is the suffix to add (default is today's date)}

\item{overwrite}{should we overwrite already existing files?}
}
\description{
When running a targets workflow, I often have files that are generated automatically
as part of the workflow. However, I often rename them to match the directory, and
add the day to the filename. This enables that automatically.
}
\details{
To rename actual targets, pass the output of \code{tar_read(target_name)}. Note that
this function will not copy and rename \code{qmd} and \code{rmd} source files.
}
