% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics_helpers.R
\name{ft_add_figure_numbers}
\alias{ft_add_figure_numbers}
\title{add numbers to figure files}
\usage{
ft_add_figure_numbers(
  figure_directory = ".",
  start_count = 1,
  count_location = "prefix",
  skip_numbers = NULL,
  regexp = "jpg$|png$|svg$"
)
}
\arguments{
\item{figure_directory}{the directory containing figure files}

\item{start_count}{what number starts the counting}

\item{count_location}{should the count be a \code{prefix} or \code{suffix}?}

\item{skip_numbers}{should any numbers be skipped due to inserting from a different location?}

\item{regexp}{what regex should be used to grab files}
}
\description{
Given that \code{rmd} and \code{qmd} chunks are \strong{normally} run sequentially, figures
should be generated in order that they are presented in the document. Therefore,
we should be able to roughly match them up with numbering generated by cross-referencing
using the modified times.
}
\seealso{
Other {Graphics}: 
\code{\link{ft_complexheatmap_scale}()},
\code{\link{ft_discretize_colorscale}()},
\code{\link{ft_ggplot2_colors}()},
\code{\link{ft_ggplot2_rotate_axis_labels}()}

Other {File Manipulation}: 
\code{\link{ft_rename_outputs}()},
\code{\link{ft_zip_directory}()}
}
\concept{{File Manipulation}}
\concept{{Graphics}}
