% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics_helpers.R
\name{ft_add_padding}
\alias{ft_add_padding}
\title{add more padding}
\usage{
ft_add_padding()
}
\value{
a string
}
\description{
Provides the code for adding padding to a plot, which is sometimes necessary
when generating UpSet plots via ComplexHeatmap::UpSet, as the labels will
be too long and get cutoff.
}
\examples{
  ft_add_padding()
}
\seealso{
Other {Graphics}: 
\code{\link{ft_add_figure_numbers}()},
\code{\link{ft_discretize_colorscale}()},
\code{\link{ft_ggplot2_add_ellipse}()},
\code{\link{ft_ggplot2_colors}()},
\code{\link{ft_ggplot2_legend_inside}()},
\code{\link{ft_ggplot2_rotate_axis_labels}()},
\code{\link{ft_ggplot2_self_colors}()},
\code{\link{ft_plot_temp_file}()},
\code{\link{ft_upset_add_counts}()},
\code{\link{ft_upset_comb_matrix}()}

Other {Code Generators}: 
\code{\link{ft_complexheatmap_scale}()},
\code{\link{ft_data_dictionary}()},
\code{\link{ft_ggplot2_add_ellipse}()},
\code{\link{ft_ggplot2_colors}()},
\code{\link{ft_ggplot2_legend_inside}()},
\code{\link{ft_ggplot2_rotate_axis_labels}()},
\code{\link{ft_ggplot2_self_colors}()},
\code{\link{ft_knitr_opts}()},
\code{\link{ft_plot_includes}()},
\code{\link{ft_remove_figures}()},
\code{\link{ft_title_block}()},
\code{\link{ft_upset_add_counts}()},
\code{\link{ft_upset_comb_matrix}()}
}
\concept{{Code Generators}}
\concept{{Graphics}}
