% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics_helpers.R
\name{ft_complexheatmap_scale}
\alias{ft_complexheatmap_scale}
\title{generate ComplexHeatmap color scale}
\usage{
ft_complexheatmap_scale()
}
\value{
string
}
\description{
The code for creating a function that maps values to colors for \code{ComplexHeatmap}
heatmaps is not easy for me to remember. Here it is.
}
\details{
My favorite colors to use include:
\itemize{
\item \code{viridis::viridis}
\item \code{scico::scico(n, palette = "vik")}
}

You can also use \code{scico::scico_palette_show()} to see other colorblind safe palettes.
}
