% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dictionary.R
\name{ft_data_dictionary}
\alias{ft_data_dictionary}
\title{template a data dictionary}
\usage{
ft_data_dictionary()
}
\value{
string
}
\description{
I always forget the syntax I use for creating a data dictionary
}
