% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dictionary.R
\name{ft_data_dictionary}
\alias{ft_data_dictionary}
\title{template a data dictionary}
\usage{
ft_data_dictionary()
}
\value{
string
}
\description{
I always forget the syntax I use for creating a data dictionary
}
\seealso{
Other {Code Generators}: 
\code{\link{ft_add_padding}()},
\code{\link{ft_complexheatmap_scale}()},
\code{\link{ft_ggplot2_add_ellipse}()},
\code{\link{ft_ggplot2_colors}()},
\code{\link{ft_ggplot2_legend_inside}()},
\code{\link{ft_ggplot2_rotate_axis_labels}()},
\code{\link{ft_ggplot2_self_colors}()},
\code{\link{ft_knitr_opts}()},
\code{\link{ft_plot_includes}()},
\code{\link{ft_remove_figures}()},
\code{\link{ft_rename_code}()},
\code{\link{ft_title_block}()},
\code{\link{ft_upset_add_counts}()},
\code{\link{ft_upset_comb_matrix}()}
}
\concept{{Code Generators}}
