% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics_helpers.R
\name{ft_discretize_colorscale}
\alias{ft_discretize_colorscale}
\title{discretize a continuous colorscale}
\usage{
ft_discretize_colorscale(
  color_scale = NULL,
  discrete_locs = c(25, 75),
  show_colors = TRUE
)
}
\arguments{
\item{color_scale}{the color scale to take colors from}

\item{discrete_locs}{where to take them from}

\item{show_colors}{create a plot showing the continuous and discrete colors}

\item{discrete_n}{how many colors to take}
}
\value{
vector of colors
}
\description{
When one is using a continuous colorscale to visualize data (say log-fold-changes or correlations),
it may be useful to create discrete colors from a usually continuous color scale.
}
\examples{
if require("scico") {
  ft_discretize_colorscale(scico::scico(100, palette = "vanimo", direction = -1),
                           discrete_locs = c(20, 80))
}

}
