% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics_helpers.R
\name{ft_ggplot2_add_ellipse}
\alias{ft_ggplot2_add_ellipse}
\title{add data ellipse}
\usage{
ft_ggplot2_add_ellipse()
}
\value{
a string
}
\description{
Adds an ellipse showing the extent of the data to a plot. Most useful for PCA
type stuff to show whether the samples separate or not.
}
\examples{
ft_ggplot2_add_ellipse()
}
\seealso{
Other {Graphics}:
\code{\link[=ft_add_figure_numbers]{ft_add_figure_numbers()}},
\code{\link[=ft_add_padding]{ft_add_padding()}},
\code{\link[=ft_discretize_colorscale]{ft_discretize_colorscale()}},
\code{\link[=ft_ggplot2_colors]{ft_ggplot2_colors()}},
\code{\link[=ft_ggplot2_legend_inside]{ft_ggplot2_legend_inside()}},
\code{\link[=ft_ggplot2_rotate_axis_labels]{ft_ggplot2_rotate_axis_labels()}},
\code{\link[=ft_ggplot2_self_colors]{ft_ggplot2_self_colors()}},
\code{\link[=ft_plot_temp_file]{ft_plot_temp_file()}},
\code{\link[=ft_upset_add_counts]{ft_upset_add_counts()}},
\code{\link[=ft_upset_comb_matrix]{ft_upset_comb_matrix()}}

Other {Code Generators}:
\code{\link[=ft_add_padding]{ft_add_padding()}},
\code{\link[=ft_complexheatmap_scale]{ft_complexheatmap_scale()}},
\code{\link[=ft_data_dictionary]{ft_data_dictionary()}},
\code{\link[=ft_ggplot2_colors]{ft_ggplot2_colors()}},
\code{\link[=ft_ggplot2_legend_inside]{ft_ggplot2_legend_inside()}},
\code{\link[=ft_ggplot2_rotate_axis_labels]{ft_ggplot2_rotate_axis_labels()}},
\code{\link[=ft_ggplot2_self_colors]{ft_ggplot2_self_colors()}},
\code{\link[=ft_knitr_opts]{ft_knitr_opts()}},
\code{\link[=ft_plot_includes]{ft_plot_includes()}},
\code{\link[=ft_remove_figures]{ft_remove_figures()}},
\code{\link[=ft_rename_code]{ft_rename_code()}},
\code{\link[=ft_title_block]{ft_title_block()}},
\code{\link[=ft_upset_add_counts]{ft_upset_add_counts()}},
\code{\link[=ft_upset_comb_matrix]{ft_upset_comb_matrix()}}

Other {ggplot2}:
\code{\link[=ft_ggplot2_colors]{ft_ggplot2_colors()}},
\code{\link[=ft_ggplot2_legend_inside]{ft_ggplot2_legend_inside()}},
\code{\link[=ft_ggplot2_rotate_axis_labels]{ft_ggplot2_rotate_axis_labels()}},
\code{\link[=ft_ggplot2_self_colors]{ft_ggplot2_self_colors()}}
}
\concept{{Code Generators}}
\concept{{Graphics}}
\concept{{ggplot2}}
