% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics_helpers.R
\name{ft_ggplot2_add_ellipse}
\alias{ft_ggplot2_add_ellipse}
\title{add data ellipse}
\usage{
ft_ggplot2_add_ellipse()
}
\value{
a string
}
\description{
Adds an ellipse showing the extent of the data to a plot. Most useful for PCA
type stuff to show whether the samples separate or not.
}
\seealso{
Other {Graphics} {Code Generators}: 
\code{\link{ft_ggplot2_legend_inside}()},
\code{\link{ft_ggplot2_self_colors}()}
}
\concept{{Graphics} {Code Generators}}
