% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics_helpers.R
\name{ft_ggplot2_colors}
\alias{ft_ggplot2_colors}
\title{use ggplot2 colors}
\usage{
ft_ggplot2_colors()
}
\value{
string
}
\description{
generates the code for getting some ggplot2 colors
}
\examples{
ft_ggplot2_colors()
}
\seealso{
Other {Graphics}: 
\code{\link{ft_add_figure_numbers}()},
\code{\link{ft_add_padding}()},
\code{\link{ft_discretize_colorscale}()},
\code{\link{ft_ggplot2_add_ellipse}()},
\code{\link{ft_ggplot2_legend_inside}()},
\code{\link{ft_ggplot2_rotate_axis_labels}()},
\code{\link{ft_ggplot2_self_colors}()},
\code{\link{ft_plot_temp_file}()},
\code{\link{ft_upset_add_counts}()},
\code{\link{ft_upset_comb_matrix}()}

Other {Code Generators}: 
\code{\link{ft_add_padding}()},
\code{\link{ft_complexheatmap_scale}()},
\code{\link{ft_data_dictionary}()},
\code{\link{ft_ggplot2_add_ellipse}()},
\code{\link{ft_ggplot2_legend_inside}()},
\code{\link{ft_ggplot2_rotate_axis_labels}()},
\code{\link{ft_ggplot2_self_colors}()},
\code{\link{ft_knitr_opts}()},
\code{\link{ft_plot_includes}()},
\code{\link{ft_remove_figures}()},
\code{\link{ft_title_block}()},
\code{\link{ft_upset_add_counts}()},
\code{\link{ft_upset_comb_matrix}()}

Other {Colors}: 
\code{\link{ft_complexheatmap_scale}()},
\code{\link{ft_discretize_colorscale}()}

Other {ggplot2}: 
\code{\link{ft_ggplot2_add_ellipse}()},
\code{\link{ft_ggplot2_legend_inside}()},
\code{\link{ft_ggplot2_rotate_axis_labels}()},
\code{\link{ft_ggplot2_self_colors}()}
}
\concept{{Code Generators}}
\concept{{Colors}}
\concept{{Graphics}}
\concept{{ggplot2}}
