% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics_helpers.R
\name{ft_ggplot2_colors}
\alias{ft_ggplot2_colors}
\title{use ggplot2 colors}
\usage{
ft_ggplot2_colors()
}
\value{
string
}
\description{
generates the code for getting some ggplot2 colors
}
\seealso{
Other {Graphics}: 
\code{\link{ft_add_figure_numbers}()},
\code{\link{ft_complexheatmap_scale}()},
\code{\link{ft_discretize_colorscale}()},
\code{\link{ft_rotate_axis_labels}()}

Other {Code Generators}: 
\code{\link{ft_complexheatmap_scale}()},
\code{\link{ft_data_dictionary}()},
\code{\link{ft_knitr_opts}()},
\code{\link{ft_plot_includes}()},
\code{\link{ft_remove_figures}()},
\code{\link{ft_rotate_axis_labels}()}
}
\concept{{Code Generators}}
\concept{{Graphics}}
