% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics_helpers.R
\name{ft_ggplot2_legend_inside}
\alias{ft_ggplot2_legend_inside}
\title{have legend inside plot area}
\usage{
ft_ggplot2_legend_inside()
}
\value{
a string
}
\description{
Provides the code necessary for putting a legend inside the plot area.
At some point the code for that changed, and I always forget it.
}
\seealso{
Other {Graphics} {Code Generators}: 
\code{\link{ft_ggplot2_add_ellipse}()},
\code{\link{ft_ggplot2_self_colors}()}
}
\concept{{Graphics} {Code Generators}}
