% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics_helpers.R
\name{ft_ggplot2_rotate_axis_labels}
\alias{ft_ggplot2_rotate_axis_labels}
\title{rotate axis labels}
\usage{
ft_ggplot2_rotate_axis_labels()
}
\value{
string
}
\description{
Gives me the code to rotate axis labels.
}
\seealso{
Other {Graphics}: 
\code{\link{ft_add_figure_numbers}()},
\code{\link{ft_complexheatmap_scale}()},
\code{\link{ft_discretize_colorscale}()},
\code{\link{ft_ggplot2_colors}()}

Other {Code Generators}: 
\code{\link{ft_complexheatmap_scale}()},
\code{\link{ft_data_dictionary}()},
\code{\link{ft_ggplot2_colors}()},
\code{\link{ft_knitr_opts}()},
\code{\link{ft_plot_includes}()},
\code{\link{ft_remove_figures}()},
\code{\link{ft_title_block}()}
}
\concept{{Code Generators}}
\concept{{Graphics}}
