% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics_helpers.R
\name{ft_ggplot2_self_colors}
\alias{ft_ggplot2_self_colors}
\title{define my own colors}
\usage{
ft_ggplot2_self_colors()
}
\value{
a string
}
\description{
Reminds me of the syntax for passing my own colors to a ggplot2 graph
}
\seealso{
Other {Graphics} {Code Generators}: 
\code{\link{ft_ggplot2_add_ellipse}()},
\code{\link{ft_ggplot2_legend_inside}()}
}
\concept{{Graphics} {Code Generators}}
