% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/md_helpers.R
\name{ft_knitr_opts}
\alias{ft_knitr_opts}
\title{knitr chunk options}
\usage{
ft_knitr_opts()
}
\value{
string
}
\description{
My preferred knitr chunk options.
}
\seealso{
Other {Markdown}: 
\code{\link{ft_plot_includes}()},
\code{\link{ft_remove_figures}()}

Other {Code Generators}: 
\code{\link{ft_complexheatmap_scale}()},
\code{\link{ft_data_dictionary}()},
\code{\link{ft_ggplot2_colors}()},
\code{\link{ft_plot_includes}()},
\code{\link{ft_remove_figures}()},
\code{\link{ft_rotate_axis_labels}()}
}
\concept{{Code Generators}}
\concept{{Markdown}}
