% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/md_helpers.R
\name{ft_plot_includes}
\alias{ft_plot_includes}
\title{plot list includes}
\usage{
ft_plot_includes()
}
\value{
string
}
\description{
Using \verb{!expr glue::glue(...)} will break \verb{quarto inspect}, which is used
by \code{targets}
}
\seealso{
Other {Markdown}:
\code{\link[=ft_knitr_opts]{ft_knitr_opts()}},
\code{\link[=ft_remove_figures]{ft_remove_figures()}},
\code{\link[=ft_title_block]{ft_title_block()}}

Other {Code Generators}:
\code{\link[=ft_add_padding]{ft_add_padding()}},
\code{\link[=ft_complexheatmap_scale]{ft_complexheatmap_scale()}},
\code{\link[=ft_data_dictionary]{ft_data_dictionary()}},
\code{\link[=ft_ggplot2_add_ellipse]{ft_ggplot2_add_ellipse()}},
\code{\link[=ft_ggplot2_colors]{ft_ggplot2_colors()}},
\code{\link[=ft_ggplot2_legend_inside]{ft_ggplot2_legend_inside()}},
\code{\link[=ft_ggplot2_rotate_axis_labels]{ft_ggplot2_rotate_axis_labels()}},
\code{\link[=ft_ggplot2_self_colors]{ft_ggplot2_self_colors()}},
\code{\link[=ft_knitr_opts]{ft_knitr_opts()}},
\code{\link[=ft_remove_figures]{ft_remove_figures()}},
\code{\link[=ft_rename_code]{ft_rename_code()}},
\code{\link[=ft_title_block]{ft_title_block()}},
\code{\link[=ft_upset_add_counts]{ft_upset_add_counts()}},
\code{\link[=ft_upset_comb_matrix]{ft_upset_comb_matrix()}}
}
\concept{{Code Generators}}
\concept{{Markdown}}
