% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/md_helpers.R
\name{ft_plot_includes}
\alias{ft_plot_includes}
\title{plot list includes}
\usage{
ft_plot_includes()
}
\value{
string
}
\description{
Using \verb{!expr glue::glue(...)} will break \verb{quarto inspect}, which is used
by \code{targets}
}
