% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knit.R
\name{ft_quarto_render}
\alias{ft_quarto_render}
\title{quarto render}
\usage{
ft_quarto_render(document = NULL, working_directory = getwd(), ...)
}
\arguments{
\item{document}{the \code{qmd} document being used}

\item{working_directory}{the root directory of the \code{targets} project (default to getwd)}

\item{...}{any other parameters used by \code{quarto::quarto_render}}
}
\value{
the output of running \code{quarto::quarto_render}
}
\description{
A version of \code{quarto::quarto_render} that has sane defaults for when working in a \code{targets} project
where report documents are frequently kept in a subfolder, but \code{targets} needs to operate from
the root folder to find and load everything.
}
\seealso{
Other 'Document Rendering': 
\code{\link{ft_rmarkdown_render}()}
}
\concept{'Document Rendering'}
