% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/md_helpers.R
\name{ft_remove_figures}
\alias{ft_remove_figures}
\title{remove figures at start}
\usage{
ft_remove_figures()
}
\value{
string
}
\description{
It happens very often, that we don't want to keep old figures laying around
from a previous rendering of the document. So this creates the code we would
insert into a setup chunk to go and delete the previously created figures.
}
\seealso{
Other {Markdown}:
\code{\link[=ft_knitr_opts]{ft_knitr_opts()}},
\code{\link[=ft_plot_includes]{ft_plot_includes()}},
\code{\link[=ft_title_block]{ft_title_block()}}

Other {Code Generators}:
\code{\link[=ft_add_padding]{ft_add_padding()}},
\code{\link[=ft_complexheatmap_scale]{ft_complexheatmap_scale()}},
\code{\link[=ft_data_dictionary]{ft_data_dictionary()}},
\code{\link[=ft_ggplot2_add_ellipse]{ft_ggplot2_add_ellipse()}},
\code{\link[=ft_ggplot2_colors]{ft_ggplot2_colors()}},
\code{\link[=ft_ggplot2_legend_inside]{ft_ggplot2_legend_inside()}},
\code{\link[=ft_ggplot2_rotate_axis_labels]{ft_ggplot2_rotate_axis_labels()}},
\code{\link[=ft_ggplot2_self_colors]{ft_ggplot2_self_colors()}},
\code{\link[=ft_knitr_opts]{ft_knitr_opts()}},
\code{\link[=ft_plot_includes]{ft_plot_includes()}},
\code{\link[=ft_rename_code]{ft_rename_code()}},
\code{\link[=ft_title_block]{ft_title_block()}},
\code{\link[=ft_upset_add_counts]{ft_upset_add_counts()}},
\code{\link[=ft_upset_comb_matrix]{ft_upset_comb_matrix()}}
}
\concept{{Code Generators}}
\concept{{Markdown}}
