% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/md_helpers.R
\name{ft_remove_figures}
\alias{ft_remove_figures}
\title{remove figures at start}
\usage{
ft_remove_figures()
}
\value{
string
}
\description{
It happens very often, that we don't want to keep old figures laying around
from a previous rendering of the document. So this creates the code we would
insert into a setup chunk to go and delete the previously created figures.
}
