% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename_outputs.R
\name{ft_rename_code}
\alias{ft_rename_code}
\title{rename code}
\usage{
ft_rename_code()
}
\value{
string
}
\description{
Provide the code I want when I rename stuff
}
\examples{
ft_rename_code()
}
\seealso{
Other {Code Generators}: 
\code{\link{ft_add_padding}()},
\code{\link{ft_complexheatmap_scale}()},
\code{\link{ft_data_dictionary}()},
\code{\link{ft_ggplot2_add_ellipse}()},
\code{\link{ft_ggplot2_colors}()},
\code{\link{ft_ggplot2_legend_inside}()},
\code{\link{ft_ggplot2_rotate_axis_labels}()},
\code{\link{ft_ggplot2_self_colors}()},
\code{\link{ft_knitr_opts}()},
\code{\link{ft_plot_includes}()},
\code{\link{ft_remove_figures}()},
\code{\link{ft_title_block}()},
\code{\link{ft_upset_add_counts}()},
\code{\link{ft_upset_comb_matrix}()}

Other {File Manipulation}: 
\code{\link{ft_add_figure_numbers}()},
\code{\link{ft_rename_outputs}()},
\code{\link{ft_zip_directory}()}
}
\concept{{Code Generators}}
\concept{{File Manipulation}}
