% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_helpers.R
\name{ft_set_function_params}
\alias{ft_set_function_params}
\title{Set function parameters to an environment.}
\source{
Jason Bryer blog, https://bryer.org/posts/2026-05-05-Setting_Function_Parameters_for_Debugging.html
}
\usage{
ft_set_function_params(FUN, envir = globalenv(), verbose = interactive())
}
\arguments{
\item{FUN}{the function to assign parameters to an environment.}

\item{envir}{the environment to assign the variables to. Defaults to the
global environment.}

\item{verbose}{whether to return the data frame invisibly or to print the results.}
}
\value{
a data frame where row names correspond to the parameter name with
two columns: \code{set} which is logical indicating if the variable was set
and \code{value} with a character representation of the variable value.
}
\description{
This function is designed to help debug functions. It will attempt to set all
the default parameter values to the specified environment (global environment
by default). This is useful for when you want to execute code within the
function definition interactively but need the parameters set in the current
environment.
}
\details{
\strong{Warning:} This function will modify the global environment and therefore
violates CRAN policy
\href{https://cran.r-project.org/web/packages/policies.html#Source-packages}{"Packages should not modify the global environment"}.
}
\seealso{
Other 'Analysis':
\code{\link[=ft_launch_watch_relaunch]{ft_launch_watch_relaunch()}},
\code{\link[=ft_limma_code]{ft_limma_code()}},
\code{\link[=ft_r_launch_watch_relaunch]{ft_r_launch_watch_relaunch()}}
}
\author{
Jason Bryer
}
\concept{'Analysis'}
