% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timekeeping.R
\name{ft_setup_timekeeping}
\alias{ft_setup_timekeeping}
\title{setup timekeeping}
\usage{
ft_setup_timekeeping()
}
\value{
text
}
\description{
setup timekeeping
}
\examples{
ft_setup_timekeeping()
}
\seealso{
Other {Timekeeping}: 
\code{\link{ft_tk_check_log}()},
\code{\link{ft_tk_get_categories}()},
\code{\link{ft_tk_logfile}()},
\code{\link{ft_tk_logtask}()},
\code{\link{ft_tk_parselog}()},
\code{\link{ft_tk_removelast}()},
\code{\link{ft_tk_summarize_time}()}
}
\concept{{Timekeeping}}
