% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/targets_progress.R
\name{ft_show_targets_progress}
\alias{ft_show_targets_progress}
\title{show targets progress}
\usage{
ft_show_targets_progress(names = NULL, ...)
}
\arguments{
\item{names}{the names of the things to track}
}
\value{
tibble
}
\description{
show targets progress
}
\seealso{
Other {targets}: 
\code{\link{ft_targets_options}()}
}
\concept{{targets}}
