% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/targets_progress.R
\name{ft_targets_options}
\alias{ft_targets_options}
\title{useful targets options}
\usage{
ft_targets_options()
}
\value{
text
}
\description{
Provides a text snippet with useful targets progress options.
}
\seealso{
Other {targets}: 
\code{\link{ft_show_targets_progress}()}
}
\concept{{targets}}
