% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/md_helpers.R
\name{ft_title_block}
\alias{ft_title_block}
\title{rmd / qmd title block}
\usage{
ft_title_block()
}
\value{
string
}
\description{
My preferred title block
}
\seealso{
Other {Markdown}: 
\code{\link{ft_knitr_opts}()},
\code{\link{ft_plot_includes}()},
\code{\link{ft_remove_figures}()}

Other {Code Generators}: 
\code{\link{ft_complexheatmap_scale}()},
\code{\link{ft_data_dictionary}()},
\code{\link{ft_ggplot2_colors}()},
\code{\link{ft_knitr_opts}()},
\code{\link{ft_plot_includes}()},
\code{\link{ft_remove_figures}()},
\code{\link{ft_rotate_axis_labels}()}
}
\concept{{Code Generators}}
\concept{{Markdown}}
