% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timekeeping.R
\name{ft_tk_logtask}
\alias{ft_tk_logtask}
\title{log starting or stopping a task}
\usage{
ft_tk_logtask(
  collab = "",
  project = "",
  category = "",
  type = "start",
  comments = "",
  log_file = ft_tk_logfile(),
  time = Sys.time()
)
}
\arguments{
\item{collab}{collaborating party}

\item{project}{the project being worked on}

\item{category}{what type of work is being done}

\item{type}{are you starting or stopping work (\code{start} or \code{stop})}

\item{comments}{futher comments about it}

\item{log_file}{where to put this info. Defaults to file returned by \code{ft_tk_logfile()}}

\item{time}{what time did this occur? Defaults to current time returned by \code{Sys.time()}}
}
\description{
Logs \strong{starting} and \strong{stopping} work on a tast to a CSV file set using the
environment variable \code{R_FT_TK_LOGFILE}. Allowed \code{categories} are set as a comma
delimited string in the environment variable \code{R_FT_TK_CATEGORIES}.
}
\seealso{
Other {Timekeeping}: 
\code{\link{ft_setup_timekeeping}()},
\code{\link{ft_tk_check_log}()},
\code{\link{ft_tk_get_categories}()},
\code{\link{ft_tk_logfile}()},
\code{\link{ft_tk_parselog}()},
\code{\link{ft_tk_removelast}()},
\code{\link{ft_tk_summarize_time}()}
}
\concept{{Timekeeping}}
