% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timekeeping.R
\name{ft_tk_logtask}
\alias{ft_tk_logtask}
\title{log starting or stopping a task}
\usage{
ft_tk_logtask(
  collab = "",
  project = "",
  category = "",
  type = "start",
  comments = "",
  log_file = ft_tk_file()
)
}
\arguments{
\item{collab}{collaborating party}

\item{project}{the project being worked on}

\item{category}{what type of work is being done}

\item{type}{are you starting or stopping work}

\item{comments}{futher comments about it}

\item{log_file}{where to put this info. Defaults to file returned by \code{ft_tk_logfile()}}
}
\description{
log starting or stopping a task
}
\seealso{
Other {Timekeeping}: 
\code{\link{ft_setup_timekeeping}()},
\code{\link{ft_tk_get_categories}()},
\code{\link{ft_tk_logfile}()},
\code{\link{ft_tk_parselog}()},
\code{\link{ft_tk_removelast}()}
}
\concept{{Timekeeping}}
