% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timekeeping.R
\name{ft_tk_removelast}
\alias{ft_tk_removelast}
\title{remove last timekeeping entry}
\usage{
ft_tk_removelast(log_file = ft_tk_file())
}
\arguments{
\item{log_file}{}
}
\value{
NULL invisibly
}
\description{
remove last timekeeping entry
}
\seealso{
Other {Timekeeping}: 
\code{\link{ft_setup_timekeeping}()},
\code{\link{ft_tk_get_categories}()},
\code{\link{ft_tk_logfile}()},
\code{\link{ft_tk_logtask}()}
}
\concept{{Timekeeping}}
