% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timekeeping.R
\name{ft_tk_summarize_time}
\alias{ft_tk_summarize_time}
\title{summarize working time}
\usage{
ft_tk_summarize_time(
  log_file = ft_tk_logfile(),
  collab_q = "",
  project_q = "",
  category_q = "",
  units = "days",
  min_cumulative = 120,
  return_type = "n-days"
)
}
\arguments{
\item{log_file}{the log file to use (ft_tk_logfile)}

\item{units}{what to summarize to? Default is "days"}

\item{min_cumulative}{minimum number of minutes to count as one "unit"}

\item{return_type}{whether to give the number of days, or specific time durations by day.}

\item{collab}{the collaborator to select}

\item{project}{the project}

\item{category}{the category of work}
}
\value{
data.frame
}
\description{
summarize working time
}
\seealso{
Other {Timekeeping}: 
\code{\link{ft_setup_timekeeping}()},
\code{\link{ft_tk_check_log}()},
\code{\link{ft_tk_get_categories}()},
\code{\link{ft_tk_logfile}()},
\code{\link{ft_tk_logtask}()},
\code{\link{ft_tk_parselog}()},
\code{\link{ft_tk_removelast}()}
}
\concept{{Timekeeping}}
