% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics_helpers.R
\name{ft_upset_comb_matrix}
\alias{ft_upset_comb_matrix}
\title{manually create combinations}
\usage{
ft_upset_comb_matrix()
}
\value{
a string
}
\description{
Provides a code framework for creating a combination matrix with code instead of using lists.
Needed because sometimes \code{ComplexHeatmap::make_comb_mat} will choke on list elements.
}
\examples{
  ft_upset_comb_matrix()
}
\seealso{
Other {Graphics}: 
\code{\link{ft_add_figure_numbers}()},
\code{\link{ft_add_padding}()},
\code{\link{ft_discretize_colorscale}()},
\code{\link{ft_ggplot2_add_ellipse}()},
\code{\link{ft_ggplot2_colors}()},
\code{\link{ft_ggplot2_legend_inside}()},
\code{\link{ft_ggplot2_rotate_axis_labels}()},
\code{\link{ft_ggplot2_self_colors}()},
\code{\link{ft_plot_temp_file}()},
\code{\link{ft_upset_add_counts}()}

Other {Code Generators}: 
\code{\link{ft_add_padding}()},
\code{\link{ft_complexheatmap_scale}()},
\code{\link{ft_data_dictionary}()},
\code{\link{ft_ggplot2_add_ellipse}()},
\code{\link{ft_ggplot2_colors}()},
\code{\link{ft_ggplot2_legend_inside}()},
\code{\link{ft_ggplot2_rotate_axis_labels}()},
\code{\link{ft_ggplot2_self_colors}()},
\code{\link{ft_knitr_opts}()},
\code{\link{ft_plot_includes}()},
\code{\link{ft_remove_figures}()},
\code{\link{ft_rename_code}()},
\code{\link{ft_title_block}()},
\code{\link{ft_upset_add_counts}()}

Other {UpSet}: 
\code{\link{ft_upset_add_counts}()}
}
\concept{{Code Generators}}
\concept{{Graphics}}
\concept{{UpSet}}
